adds support for Aqua

+++ src/sortedintset.jl
@@ -10,7 +10,7 @@ end
 @inline Base.size(s::SortedIntSet) = size(s.set)
 @inline Base.length(s::SortedIntSet) = length(s.set)
 @inline Base.eachindex(s::SortedIntSet) = eachindex(s.set)
-@inline Base.getindex(s::SortedIntSet, i) = @inbounds s.set[i]
-@inline Intersections.getkey(s::SortedIntSet, i) = @inbounds s.set[i]
+@inline Base.getindex(s::SortedIntSet, i::Integer) = @inbounds s.set[i]
+@inline Intersections.getkey(s::SortedIntSet, i::Integer) = @inbounds s.set[i]
 @inline Base.first(s::SortedIntSet) = @inbounds s[1]
 @inline Base.last(s::SortedIntSet) = @inbounds s[end]
